VCST-4675: Removed repository from controller
