+++ src/VirtoCommerce.CatalogModule.Core/Services/ICategoryIndexingService.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace VirtoCommerce.CatalogModule.Core.Services
+{
+    public interface ICategoryIndexingService
+    {
+        Task<IList<string>> GetProductIdsForIndexAsync(string categoryId);
+    }
+}
