+++ src/VirtoCommerce.CatalogModule.Web/Module.cs
@@ -123,6 +123,7 @@ namespace VirtoCommerce.CatalogModule.Web
             serviceCollection.AddTransient<IListEntrySearchService, ListEntrySearchService>();
 
             serviceCollection.AddTransient<ICategoryService, CategoryService>();
+            serviceCollection.AddTransient<ICategoryIndexingService, CategoryIndexingService>();
             serviceCollection.AddTransient<ICategoryIndexedSearchService, CategoryIndexedSearchService>();
 
             serviceCollection.AddTransient<IItemService, ItemService>();
